+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Fixed a JavaScript error that occurred when filtering out a Neo block type in a case where the block type was already in use; now, Neo blocks with filtered-out types won't be rendered, and will be deleted if structural changes are made to the Neo field
 - Fixed a JavaScript error that occurred as a result of cloning a block that uses the child blocks UI element
 - Fixed a bug where disabled blocks weren't displaying on entry revisions
+- Fixed a bug where block structures weren't being resaved when saving a provisional draft
 
 ## 3.1.1 - 2022-06-19